Refactor DremioFlightProducer so that it implements FlightSqlProducer, remova DremioFlightSqlProducer
